+++ src/js_user.hpp
@@ -93,12 +93,14 @@ public:
     static void session_for_on_disk_path(ContextType, ObjectType, Arguments&, ReturnValue&);
     static void delete_user(ContextType, ObjectType, Arguments&, ReturnValue&);
     static void link_credentials(ContextType, ObjectType, Arguments&, ReturnValue&);
+    static void call_function(ContextType, ObjectType, Arguments&, ReturnValue&);
 
     MethodMap<T> const methods = {
         {"logOut", wrap<logout>},
         {"_sessionForOnDiskPath", wrap<session_for_on_disk_path>},
         {"_deleteUser", wrap<delete_user>},
-        {"_linkCredentials", wrap<link_credentials>}
+        {"_linkCredentials", wrap<link_credentials>},
+        {"_callFunction", wrap<call_function>},
     };
 };
 
@@ -254,7 +256,8 @@ void UserClass<T>::link_credentials(ContextType ctx, ObjectType this_object, Arg
         }
 
         ValueType callback_arguments[2];
-        callback_arguments[0] = create_object<T, UserClass<T>>(ctx, new User<T>(std::move(shared_user), std::move(user->m_app)));
+        callback_arguments[0] = create_object<T, UserClass<T>>(protected_ctx, new User<T>(std::move(shared_user),
+                                                                                          user->m_app));
         callback_arguments[1] = Value::from_undefined(protected_ctx);
         Function::callback(protected_ctx, protected_callback, typename T::Object(), 2, callback_arguments);
     });
@@ -262,6 +265,35 @@ void UserClass<T>::link_credentials(ContextType ctx, ObjectType this_object, Arg
     user->m_app->link_user(*user, credentials, callback_handler);
 }
 
+template<typename T>
+void UserClass<T>::call_function(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue &) {
+    args.validate_count(3);
+    auto user = get_internal<T, UserClass<T>>(ctx, this_object);
+
+    auto name = Value::validated_to_string(ctx, args[0], "name");
+    auto call_args_js = Value::validated_to_array(ctx, args[1], "args");
+    auto callback = Value::validated_to_function(ctx, args[2], "callback");
+    auto call_args_bson = Value::to_bson(ctx, call_args_js);
+
+    user->m_app->call_function(
+        *user,
+        name,
+        call_args_bson.operator const bson::BsonArray&(), 
+        EventLoopDispatcher([ctx = Protected(Context<T>::get_global_context(ctx)),
+                             callback = Protected(ctx, callback),
+                             this_object = Protected(ctx, this_object)]
+                            (util::Optional<app::AppError> error, util::Optional<bson::Bson> result) {
+            HANDLESCOPE(ctx);
+            // Note: reversing argument order.
+            Function::callback(ctx, callback, this_object, {
+                !result ? Value::from_undefined(ctx) : Value::from_bson(ctx, *result),
+                !error ? Value::from_undefined(ctx) : Object::create_obj(ctx, {
+                    {"message", Value::from_string(ctx, error->message)},
+                    {"code", Value::from_number(ctx, error->error_code.value())},
+                }),
+            });
+        }));
+}
 
 }
 }